test: refactor tests into testUtils folder
